requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
APScheduler==3.6.3
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
bleach==3.1.5
boto3==1.13.15
botocore==1.16.15
bz2file==0.98
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==1.2.3
contextvars==2.4
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dopamine-rl==3.0.1
fbmessenger==6.0.0
Flask==1.1.2
future==0.18.2
gast==0.2.2
gevent==1.5.0
gin-config==0.3.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-pasta==0.2.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.29.0
gunicorn==20.0.4
gym==0.15.4
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.5.19
httpcore==0.3.0
httplib2==0.18.1
httptools==0.1.1
httpx==0.11.1
humanfriendly==8.2
hyperframe==5.2.0
idna==2.9
idna-ssl==1.1.0
immutables==0.14
importlib-metadata==1.6.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.10.0
jsonpickle==1.4.1
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==21.2.1
kfac==0.2.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mattermostwrapper==2.2
mesh-tensorflow==0.1.13
mitie==0.7.0
mpmath==1.1.0
multidict==4.6.1
neobolt==1.7.17
neotime==1.7.4
networkx==2.4
numpy==1.18.4
oauth2client==4.1.3
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==19.2
pandas==1.0.3
pika==1.0.1
Pillow==7.1.2
pkginfo==1.5.0.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.12.1
py2neo==4.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
pyglet==1.3.2
Pygments==2.6.1
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
pypng==0.0.20
pyreadline==2.1
pyrsistent==0.16.0
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
python-telegram-bot==11.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
questionary==1.5.2
rasa==1.7.4
rasa-sdk==1.7.0
readme-renderer==26.0
redis==3.3.11
requests==2.23.0
requests-async==0.5.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==0.6.36
rsa==4.0
ruamel.yaml==0.15.100
s3transfer==0.3.3
sanic==19.9.0
Sanic-Cors==0.9.9.post1
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.2
scikit-learn==0.20.4
scipy==1.4.1
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
sniffio==1.1.0
SQLAlchemy==1.3.17
sympy==1.5.1
tabulate==0.8.7
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
tensorflow-datasets==3.1.0
tensorflow-gan==2.0.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.46.0
treelib==1.6.1
twilio==6.40.0
twine==3.1.0
typing-extensions==3.7.4.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
webencodings==0.5.1
webexteamssdk==1.3
websocket-client==0.54.0
websockets==8.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
yarl==1.4.2
zipp==3.1.0
